# Improve jq_warm_thrusters performance 