As discussed in IRC: We wanted to explicitly whitelist about 8000 devices in a larger network using a script-generated devices_only, but the mgr: init phase then took over an hour with one CPU maxed out.

The problem at heart is that  _get_denied_actions will do very expensive check_acl calculations on the full devices_only list for every device and every list derived from it (like discover_only etc.).

Here the currently released Netdisco with some prints in JobQueue::PostgreSQL:

```
netdisco-daemon-fg
[26889] 2019-04-24 15:16:55  warn App::Netdisco 2.042006 backend
[26891] 2019-04-24 15:16:56  warn jq_warm_thrusters: 0 of 8052 devices took 0.500218868255615 seconds
[26891] 2019-04-24 15:17:07  warn jq_warm_thrusters: 20 of 8052 devices took 11.4410638809204 seconds
[26891] 2019-04-24 15:17:18  warn jq_warm_thrusters: 40 of 8052 devices took 22.3713128566742 seconds
[26891] 2019-04-24 15:17:30  warn jq_warm_thrusters: 60 of 8052 devices took 34.4152607917786 seconds
[26891] 2019-04-24 15:17:41  warn jq_warm_thrusters: 80 of 8052 devices took 45.4804749488831 seconds
[26891] 2019-04-24 15:17:51  warn jq_warm_thrusters: 100 of 8052 devices took 55.1120328903198 seconds
```

This PR adds a configuration option `memoize_acl: (defaults to false)` which will install a memoized version of check_acl, as well as cache the expensive `NetAddr::IP::Lite` constructor (you can see in this [NYTProf flamegraph of the unpatched code](https://www.networkz.ch/nytprof_manager_init/) that it's the biggest factor by far). The former greatly speeds up checking entries from duplicated devices_only/no lists, while the latter helps with speeding up check_acl in general.

Same configuration with the patched Permission.pm:

```
netdisco-daemon-fg
[30535] 2019-04-24 15:24:35  warn App::Netdisco 2.042006 backend
[30537] 2019-04-24 15:24:36  warn jq_warm_thrusters: 0 of 8052 devices took 0.147835969924927 seconds
[30537] 2019-04-24 15:24:38  warn jq_warm_thrusters: 20 of 8052 devices took 1.91254997253418 seconds
[30537] 2019-04-24 15:24:40  warn jq_warm_thrusters: 40 of 8052 devices took 3.60210299491882 seconds
[30537] 2019-04-24 15:24:41  warn jq_warm_thrusters: 60 of 8052 devices took 5.49756407737732 seconds
[30537] 2019-04-24 15:24:43  warn jq_warm_thrusters: 80 of 8052 devices took 7.30288600921631 seconds
[30537] 2019-04-24 15:24:45  warn jq_warm_thrusters: 100 of 8052 devices took 8.88906693458557 seconds
```

Actually I don't have a lot of experience with Perl memoization and the fancy Moo stuff we're using here, so I don't just want to merge it without some second opinions. Would you guys be comfortable with this code, giving it's disabled by default and only users with long ACLs could enable it on a as-needed basis?

Cheers
Christian