@ollyg 2019-04-28
I've implemented string compare in:
[master ca74f32e] improve check_acl performance for basic IP string compare